api_yamdb/requirements.txt
requests==2.26.0
Django==3.2
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==22.1
django-import-export==3.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4
psycopg2-binary==2.9.9
pytz==2020.1
sqlparse==0.3.1
